+++ cli/cli/kurtosis_config/v1/kubernetes_cluster_config_v1.go
@@ -1,0 +1,8 @@
+package v1
+
+type KubernetesClusterConfigV1 struct {
+	KubernetesClusterName *string `yaml:"kubernetes-cluster-name,omitempty"`
+	StorageClass *string `yaml:"storage-class,omitempty"`
+	EnclaveSizeInGigabytes *uint `yaml:"enclave-size-in-gigabytes,omitempty"`
+}
+